+++ tests/unit_tests/backtest/test_backtest_config.py
@@ -47,7 +47,7 @@ from tests.test_kit.stubs.persistence import TestPersistenceStubs
 
 TEST_DATA_DIR = str(pathlib.Path(PACKAGE_ROOT).joinpath("data"))
 
-pytestmark = pytest.mark.skipif(sys.platform == "win32", reason="test path broken on windows")
+pytestmark = pytest.mark.skipif(sys.platform == "win32", reason="test path broken on Windows")
 
 
 @dataclasses.dataclass(repr=False)